#1002: Continue working to separate jrunscript/tools/install deprecated items to separate script file
